layout rendering bug fixed
